feat: add options of bug and feature request to dropdown menu
