+++ nodes/web/bug_report.js
@@ -1,0 +1,21 @@
+import { app } from "../../scripts/app.js";
+import { style } from "./helpers/bug_report/main.js";
+import { createMenuItem, findMenuList } from "./helpers/global/main.js";
+
+// Add the CSS class for the bug icon
+document.head.appendChild(style);
+
+const ext = {
+  name: "signature.bug_report",
+  async init(app) {
+    const menuList = findMenuList();
+    if (menuList) {
+      menuList.appendChild(
+        createMenuItem("Bug Report", "bug-report-icon", () => {
+          window.open("https://signature-ai.atlassian.net/jira/software/c/projects/SIGML/form/69", "_blank");
+        })
+      );
+    }
+  },
+};
+app.registerExtension(ext);
