+++ nodes/web/helpers/bug_report/main.js
@@ -1,0 +1,2 @@
+import { style } from "./style.js";
+export { style };
