+++ nodes/web/helpers/bug_report/style.js
@@ -1,0 +1,23 @@
+const style = document.createElement("style");
+style.textContent = `
+  .bug-report-icon::before {
+    content: '';
+    display: inline-block;
+    width: 22px;
+    height: 22px;
+    position: relative;
+    left: -3px;
+    background-image: url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" width="22" height="22" viewBox="0 0 22 22"><path d="M20 8h-2.81c-.45-.78-1.07-1.45-1.82-1.96L17 4.41 15.59 3l-2.17 2.17C12.96 5.06 12.49 5 12 5s-.96.06-1.41.17L8.41 3 7 4.41l1.62 1.63C7.88 6.55 7.26 7.22 6.81 8H4v2h2.09c-.05.33-.09.66-.09 1v1H4v2h2v1c0 .34.04.67.09 1H4v2h2.81c1.04 1.79 2.97 3 5.19 3s4.15-1.21 5.19-3H20v-2h-2.09c.05-.33.09-.66.09-1v-1h2v-2h-2v-1c0-.34-.04-.67-.09-1H20V8zm-4 4v3c0 2.21-1.79 4-4 4s-4-1.79-4-4v-3c0-2.21 1.79-4 4-4s4 1.79 4 4z" fill="rgb(113, 113, 122)"/></svg>');
+    background-size: contain;
+    background-repeat: no-repeat;
+    vertical-align: middle;
+    filter: brightness(0.8);  /* Slightly dim the icon to match others */
+  }
+
+  .bug-report-icon {
+    display: flex;
+    align-items: center;
+    color: #717171;
+  }`;
+
+export { style };
