+++ nodes/web/helpers/workflow/utils.js
@@ -10,6 +10,8 @@ import { saveWorkflow } from "./form/main.js";
 import { showWorkflowsList } from "./list/main.js";
 import { showNodeOrderEditor } from "./node_order/main.js";
 
+import { $el, createMenuItem, findMenuList } from "../global/main.js";
+
 const showMessage = (message, color, detailedInfo = null, backgroundColor = "#00000000", extraBody = null) => {
   let dialogContent = `
         <div style="
@@ -90,21 +92,6 @@ const getLoadingSpinner = (color) => {
     `;
 };
 
-const findMenuList = () => {
-  // Try different possible menu list IDs
-  const possibleMenuLists = [
-    "#pv_id_9_0_list",
-    "#pv_id_10_0_list",
-    ".p-menubar-root-list", // Backup selector
-  ];
-
-  for (const selector of possibleMenuLists) {
-    const menuList = document.querySelector(selector);
-    if (menuList) return menuList;
-  }
-  return null;
-};
-
 const deleteWorkflowFromStorage = (tabIndex) => {
   // Delete the workflow at the current index
   const currentKey = `workflow - ${tabIndex}`;
@@ -126,46 +113,6 @@ const deleteWorkflowFromStorage = (tabIndex) => {
     nextIndex++;
     nextKey = `workflow - ${nextIndex}`;
   }
-};
-
-const $el = (tag, propsOrChildren, children) => {
-  const split = tag.split(".");
-  const element = document.createElement(split.shift());
-  if (split.length > 0) {
-    element.classList.add(...split);
-  }
-  if (propsOrChildren) {
-    if (typeof propsOrChildren === "string") {
-      propsOrChildren = { textContent: propsOrChildren };
-    } else if (propsOrChildren instanceof Element) {
-      propsOrChildren = [propsOrChildren];
-    }
-    if (Array.isArray(propsOrChildren)) {
-      element.append(...propsOrChildren);
-    } else {
-      const { parent, $: cb, dataset, style, ...rest } = propsOrChildren;
-      if (rest.for) {
-        element.setAttribute("for", rest.for);
-      }
-      if (style) {
-        Object.assign(element.style, style);
-      }
-      if (dataset) {
-        Object.assign(element.dataset, dataset);
-      }
-      Object.assign(element, rest);
-      if (children) {
-        element.append(...(Array.isArray(children) ? children : [children]));
-      }
-      if (parent) {
-        parent.append(element);
-      }
-      if (cb) {
-        cb(element);
-      }
-    }
-  }
-  return element;
 };
 
 const cleanLocalStorage = () => {
@@ -331,38 +278,6 @@ const showLoginForm = () => {
 
   app.ui.dialog.show(formContent);
   return formContent;
-};
-
-const createMenuItem = (label, iconClass, onClick) => {
-  const menuItem = $el(
-    "li",
-    {
-      className: "p-menubar-item relative",
-      role: "menuitem",
-      "aria-label": label,
-    },
-    [
-      $el("div", { className: "p-menubar-item-content" }, [
-        $el(
-          "a",
-          {
-            className: "p-menubar-item-link",
-            "data-v-6fecd137": "",
-            onclick: (event) => {
-              event.preventDefault();
-              event.stopPropagation();
-              onClick();
-            },
-          },
-          [
-            $el("span", { className: `p-menubar-item-icon pi ${iconClass}` }),
-            $el("span", { className: "p-menubar-item-label", textContent: label }),
-          ]
-        ),
-      ]),
-    ]
-  );
-  return menuItem;
 };
 
 const requiresAuth = async (app, next) => {
